+++ install/install.js
@@ -55,7 +55,7 @@ function CheckPage()
             error = "Fields need to be filled in!";
         } else {
             data = '{"root_path":"'+$("#root_path").val()+'", "url_path":"'+$("#url_path").val()+'"}';
-            tasks = ["folder*install", "folder*includes", "folder*files", "folder*upload", "extension*mcrypt", "extension*mbstring", "extension*openssl", "extension*bcmath", "extension*iconv", "function*mysqli_fetch_all", "version*php", "ini*max_execution_time", "folder*includes/avatars", "extension*xml"];
+            tasks = ["folder*install", "folder*includes", "folder*files", "folder*upload", "extension*mcrypt", "extension*mbstring", "extension*openssl", "extension*bcmath", "extension*iconv", "extension*gd", "function*mysqli_fetch_all", "version*php", "ini*max_execution_time", "folder*includes/avatars", "extension*xml"];
             multiple = true;
             $("#hid_abspath").val($("#root_path").val());
             $("#hid_url_path").val($("#url_path").val());
@@ -142,7 +142,12 @@ function CheckPage()
                     } else {
                         data = $.parseJSON(data.responseText);
                         if (data[0].error == "") {
-                            $("#res"+step+"_check"+data[0].index).html("<img src=\"images/tick.png\">");
+                            if (step == "5") {
+                                $('#pop_db').append('<li>Table <b>'+data[0].table+'</b> created</li>');
+                            } else {
+                                $("#res"+step+"_check"+data[0].index).html("<img src=\"images/tick.png\">");
+                            }
+                            
                             if (data[0].result != undefined && data[0].result != "" ) {
                                 $("#step_result").html(data[0].result);
                             }