+++ sources/users.queries.php
@@ -455,6 +455,7 @@ if (!empty($_POST['type'])) {
                 "id = %i",
                 $_POST['id']
             );
+            echo prepareExchangedData(array("error" => ""), "encode");
             break;
 
         /**